Added fzf and xtrackto

+++ setup/setup.sh
@@ -2,10 +2,8 @@
 CURDIR="$(dirname $(readlink -f "$0"))"
 find "$CURDIR/.." -maxdepth 1 -not -iname setup -not -iname .. -not -iname . -not -iname .git | while read each ; do
     DEST="$HOME/$(basename "$each")"
-    if [ -e "$DEST" ] ; then
-        mv -v "$DEST" "$DEST.bak" || echo "$DEST already exists."
-    fi
-    ln -v -s "$each" "$DEST"
+    [ -e "$DEST" ] && mv -v "$DEST" "$DEST.bak" || echo "$DEST already exists."
+    [ ! -e "$DEST" ] && ln -v -s "$each" "$DEST"
 done
 
 (
@@ -14,3 +12,5 @@ done
 )
 
 pip3 install --upgrade neovim
+sudo npm install -g tern
+"$CURDIR/../.vim/plugged/fzf/install"
